requirements.txt
scikit-learn
apache-airflow[postgres,redis,celery]
psycopg2-binary
requests
pandas
numpy
apache-airflow-providers-amazon
apache-airflow-providers-http
apache-airflow-providers-docker
apache-airflow-providers-slack
apache-airflow-providers-postgres